feat(audit-remediation): plan-2026-07-29-2005 R1.1+R1.2 ORM 机械模型一致性修复批次 1
- R1.1: 5 域受影响实体 propId 重编号连续无缺口（mfg/assets/projects/maintenance/quality）
- R1.2: crm 7 列 stdDataType double→decimal + 6 处调用方 BigDecimal 适配（ForecastAggregator/FunnelAggregationEngine/PriceRuleEngine/2 BizModel/2 测试类）
- codegen 增量再生（_gen 实体 BigDecimal + _app.orm.xml + xmeta + view.xml + i18n + InputBean/OutputBean + deploy sql）

Plan: docs/plans/2026-07-29-2005-1-orm-ask-first-mechanical-model-conformance-batch1.md (full-green verification)

+++ module-crm/erp-crm-service/src/main/java/app/erp/crm/service/entity/ErpCrmLeadFunnelBizModel.java
@@ -11,6 +11,8 @@ import app.erp.crm.dao.entity.ErpCrmLostReason;
 import app.erp.crm.dao.entity.ErpCrmStage;
 import app.erp.crm.service.ErpCrmConfigs;
 import app.erp.crm.service.support.FunnelAggregationEngine;
+
+import java.math.BigDecimal;
 import io.nop.api.core.annotations.biz.BizModel;
 import io.nop.api.core.annotations.biz.BizMutation;
 import io.nop.api.core.annotations.biz.BizQuery;
@@ -104,7 +106,7 @@ public class ErpCrmLeadFunnelBizModel extends CrudBizModel<ErpCrmLeadFunnel> imp
         funnel.setLostRevenue(snapshot.getLostRevenue());
         funnel.setWeightedRevenue(snapshot.getWeightedRevenue());
         funnel.setAvgDealSize(snapshot.getAvgDealSize());
-        funnel.setAvgSalesCycleDays(snapshot.getAvgSalesCycleDays());
+        funnel.setAvgSalesCycleDays(BigDecimal.valueOf(snapshot.getAvgSalesCycleDays()));
         funnel.setCalculatedAt(snapshot.getCalculatedAt() != null ? Timestamp.valueOf(snapshot.getCalculatedAt()) : null);
         saveEntity(funnel, null, context);
